d/salsa-ci.yml: Run extract-source job in target release, not unstable
Currently the extract-source job is run on a container image that is
always based on unstable, not the target release.  Adding sources for
the target release should make CI builds work but it's not consistent
with the way we normally prepapre source packages.

Change the image to be $SALSA_CI_IMAGES_BASE and install the extra
packages that are needed.

(cherry picked from commit 22c0ba75f4c447e84e76e3bab06f31d920f677e1)