d/b/buildcheck.py: Reimplement check on kernel file size
Add a config field for maximum kernel file size, and a checker for
this in buildcheck.py.  This is still needed for sh4/sh7785lcr.  We no
longer have any flavours with a limit on the uncompressed kernel size
or the size with a DTB appended, so don't reimplement those.

This partially reverts commit b1a1d357d647 "Remove unused check for
image size".

Fixes: a7a67cc5b947 ("Use new TOML config")