[arm64] Add {of,nd}_pmem to arm64 md-modules udeb
U-Boot is adding support for HTTP(s) boot. The installation media is made
available as /dev/pmem: in order for the installer to be able to mount it, we
need to have pmem modules in the installer initrd.

This is the first MR needed to support U-Boot HTTP(s) boot, the other two being:

- Add md-modules to arm64 cdrom builds
- Add /dev/pmem devices to list-devices-linux in debian-installer-utils

See https://www.linaro.org/blog/ledge-blogs-uefi-http-and-https-boot-in-u-boot/