d/b/gencontrol.py: Fix generation of debian/tests/control
Currently we fail to include headers-kbuild tests in
debian/tests/control.

We read the template with read_deb822() which is a generator, but we
try to iterate over it twice which results in a an empty sequence the
second time.

Capture the generator output into a list so this works again.

Fixes: 1f3a3d27318a ("Convert control entries into dataclasses")